minqx-main (1)/minqx-main/requirements.txt
python-telegram-bot==20.3
SQLAlchemy==1.4.46
psycopg2-binary==2.9.6
Flask==2.2.2
